Tests/LibWeb/Text/expected/SVG/svg-foreignObject-abspos-preserved-during-partial-relayout.txt

Before SVG relayout:
  abspos width: 25
  abspos height: 25
  abspos x: 18
  abspos y: 18
After SVG relayout:
  abspos width: 25
  abspos height: 25
  abspos x: 18
  abspos y: 18
Abspos preserved after SVG relayout:
  width unchanged: true
  height unchanged: true
  x unchanged: true
  y unchanged: true
